Undo changes to Payment entity, and add PaymentSession entity
- Undo changes previously made to the Payment entity and db table.
These changes were adding 'finishedAt' and 'status' columns, and
renaming the 'madeAt' column to 'startedAt'. The changes were made to
allow the Payment entity to be (partly) a session, ie, the session
from when a payment has started to when it is completed (successfully
or unsuccessfully). All these changes have been undone and the new
PaymentSession entity will be used to represent a payment session.
- Add a PaymentSession entity and db table to represent and carry
data about a payment session. A payment session starts when a
payment process is started to when it is completed, successfully
or unsuccessfully.)